Extract components from DashboardHeader into their own file.
