--- frontend/dashboard/pages/PageLayout/DashboardHeader/DashboardHeader.module.css
@@ -1,3 +1,0 @@
-.active {
-  border-bottom: 2px solid var(--fds-semantic-surface-neutral-default);
-}
