+++ frontend/dashboard/pages/PageLayout/DashboardHeader/LargeNavigationMenu/LargeNavigationMenu.module.css
@@ -1,0 +1,14 @@
+.menu {
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  list-style-type: none;
+  flex: 2;
+  padding: 0 var(--fds-spacing-4);
+  gap: var(--fds-spacing-2);
+  font-size: var(--fds-sizing-4);
+}
+
+.active {
+  border-bottom: 2px solid var(--fds-semantic-surface-neutral-default);
+}
