+++ frontend/dashboard/pages/PageLayout/DashboardHeader/LargeNavigationMenu/index.ts
@@ -1,0 +1,2 @@
+export { LargeNavigationMenu } from './LargeNavigationMenu';
+export type { LargeNavigationMenuProps } from './LargeNavigationMenu';
